+++ src/main/java/kang/tableorder/repository/CustomerReviewRepository.java
@@ -11,9 +11,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface CustomerReviewRepository extends JpaRepository<CustomerReviewEntity, Integer> {
+public interface CustomerReviewRepository extends JpaRepository<CustomerReviewEntity, Long> {
 
-  Optional<CustomerReviewEntity> findByIdAndUserEntity(Integer id, UserEntity userEntity);
+  Optional<CustomerReviewEntity> findByIdAndUserEntity(Long id, UserEntity userEntity);
 
   @Modifying
   @Query("DELETE FROM CUSTOMER_REVIEW cr "
@@ -21,7 +21,7 @@ public interface CustomerReviewRepository extends JpaRepository<CustomerReviewEn
       + "AND cr.restaurantEntity.id = :restaurantId "
       + "AND cr.menuEntity.id = :menuId "
       + "AND cr.userEntity = :userEntity")
-  void deleteByAllId(Integer id, Integer restaurantId, Integer menuId, UserEntity userEntity);
+  void deleteByAllId(Long id, Long restaurantId, Long menuId, UserEntity userEntity);
 
   List<CustomerReviewEntity> findAllByUserEntity(UserEntity userEntity);
 
@@ -29,13 +29,13 @@ public interface CustomerReviewRepository extends JpaRepository<CustomerReviewEn
 
   Page<CustomerReviewEntity> findAllByMenuEntity(MenuEntity menuEntity, Pageable pageable);
 
-  Optional<CustomerReviewEntity> findByIdAndMenuEntity(Integer id, MenuEntity menuEntity);
+  Optional<CustomerReviewEntity> findByIdAndMenuEntity(Long id, MenuEntity menuEntity);
 
   @Query("SELECT c FROM CUSTOMER_REVIEW c "
       + "WHERE c.id = :id "
       + "AND c.restaurantEntity.id = :restaurantId "
       + "AND c.menuEntity.id = :menuId "
       + "AND c.userEntity = :userEntity")
-  Optional<CustomerReviewEntity> findByAllId(Integer id, Integer restaurantId, Integer menuId,
+  Optional<CustomerReviewEntity> findByAllId(Long id, Long restaurantId, Long menuId,
       UserEntity userEntity);
 }
